make a good mutation
