+++ packages/app/src/pages/Settings.vue
@@ -1,18 +1,12 @@
 <template>
-  <div>
-    <Button @click="reconfigure">
-      Reconfigure
-    </Button>
-    <SettingsContainer
-      v-if="query.data.value"
-      :gql="query.data.value"
-    />
-  </div>
+  <SettingsContainer
+    v-if="query.data.value"
+    :gql="query.data.value"
+  />
 </template>
 
 <script lang="ts" setup>
 import { gql, useQuery, useMutation } from '@urql/vue'
-import Button from '@cy/components/Button.vue'
 import SettingsContainer from '../settings/SettingsContainer.vue'
 import { Settings_ReconfigureProjectDocument, SettingsDocument } from '../generated/graphql'
